This current version should produce a results file that has times in the first column, then mean female worm burdens for each individual run in the following columns. It crashes right now when executed so need to identify the source of the crash (likely to do with code in the CRealization.cpp file). Speak to James about this issue.

+++ CWorm.h
@@ -14,7 +14,7 @@ public:
 	virtual ~CWorm();
 
 	// Members.
-	int femaleWorms, totalWorms;
+	double femaleWorms, totalWorms;
 };
 
 #endif /* CWORM_H_ */
